requirements.txt
flask
flask_restful
flask_sqlalchemy
requests
attrs